v2/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
alembic
pydantic
spacy
rapidfuzz
sentence-transformers
faiss-cpu
python-dotenv
pytest
